// src/types/setting.ts
export type SettingInputTypes = {
    email: string;
    password: string;
    emailNotifications: boolean;
    monthlyReport: boolean;
    promotionalInfo: boolean;
  };

export type SupportInputTypes = {
    subject: string;
    message: string;
  };

export type FeatureInputTypes = {
    message: string;
  };